Fixed issue
in RC4 it generates a bit stream based on the key. If we reset the bit
stream, we're basically doing XOR

RC4(RC4(0xAABBCCDD, 0x11223344), 0x11223344)

should be equal to:

XOR(XOR(0xAABBCCDD, 0x11223344), 0x11223344)

assuming the first bytes of the bit stream are equal to the key, which
they may not be. Going to test this.